scripts/requirements.txt
matplotlib==3.5.1
psycopg2==2.9.3
requests==2.26.0
schedule==1.1.0
numpy<2
python-dotenv>=0.19.0

